requirements.txt
Flask==3.1.0
Flask_JWT_Extended==4.7.1
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.1
